requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
python-decouple>=3.8
pandas>=2.0.0
openpyxl>=3.1.0
psycopg2-binary>=2.9.0

